requirements.txt
python-dotenv
psycopg2-binary
click
boto3